chore: extract resource actions from resource list to separate file
