+++ packages/frontend/src/components/common/ResourceList/ResourceTable.tsx
@@ -3,8 +3,9 @@ import { Tooltip2 } from '@blueprintjs/popover2';
 import { assertUnreachable } from '@lightdash/common';
 import React, { FC, useMemo, useState } from 'react';
 import { useHistory, useParams } from 'react-router-dom';
-import { ResourceListActionState, ResourceListCommonProps } from '.';
+import { ResourceListCommonProps } from '.';
 import { useSpaces } from '../../../hooks/useSpaces';
+import { ResourceListActionState } from './ResourceActionHandlers';
 import ResourceActionMenu from './ResourceActionMenu';
 import ResourceIcon from './ResourceIcon';
 import ResourceLastEdited from './ResourceLastEdited';
